adding get node version (#4)
Adding a handy workflow so we can re-use it among different `nodejs` based repositories when running workflows. 

a typical use of this workflow would look like

let's say we wish to run linting on a given package:

```yml

name: Lint

on:
  push:
    branches:
      - main
  pull_request:

jobs:
  getNodeVersion:
    uses: calyptia/.github/workflow_templates/get_node_version.yml@main

  lint:
    runs-on: ubuntu-latest
    needs: getNodeVersion
    steps:
      - uses: actions/checkout@main
      - name: Setting node version to ${{ needs.getNodeVersion.outputs.node_version }}
        uses: actions/setup-node@v2
        with:
          node-version: '${{ needs.getNodeVersion.outputs.node_version }}'
      - run: npm install
      - run: npm run lint
```
I've also added events to fluent-linter.yml workflow to make it trigger on push as well.